requirements.txt
streamlit
streamlit-webrtc
language-tool-python
pydub
numpy
opencv-python
sounddevice
soundfile
requests
firebase-admin
supabase
fpdf2
reportlab
openai-whisper
google-generativeai

